chore: create shared prisma file
