+++ src/scheduler/batch/tasks/updateListsForAnnualReview.ts
@@ -1,4 +1,4 @@
-import { findListByAnnualReviewDate, updateListForAnnualReview } from "server/models/list";
+import { findListByAnnualReviewDate, updateListForAnnualReview } from "shared/listHelpers";
 import type { List } from "server/models/types";
 import { logger } from "scheduler/logger";
 import { findListItems } from "server/models/listItem";
@@ -55,7 +55,7 @@ export async function populateCurrentAnnualReview(lists: List[]): Promise<void> 
       }
 
       if (isUpdateList) {
-        await updateListForAnnualReview(list, { currentAnnualReview });
+        await updateListForAnnualReview(list, { currentAnnualReview }, logger);
       }
     }
   }
@@ -64,7 +64,7 @@ export async function populateCurrentAnnualReview(lists: List[]): Promise<void> 
 export async function updateListsForAnnualReview(today: Date): Promise<void> {
   const annualReviewStartDate = addDays(today, schedulerMilestoneDays.post.ONE_MONTH);
   if (annualReviewStartDate) {
-    const { result: lists } = await findListByAnnualReviewDate(annualReviewStartDate);
+    const { result: lists } = await findListByAnnualReviewDate(annualReviewStartDate, logger);
     logger.info(
       `Found the lists ${lists?.map((list) => list.id)} matching annual review start date [${annualReviewStartDate}]`
     );
