+++ src/scheduler/workers/unpublish/findListsInAnnualReview.ts
@@ -1,5 +1,5 @@
 import { startOfToday } from "date-fns";
-import { prisma } from "server/models/db/prisma-client";
+import { prisma } from "shared/prisma";;
 
 export async function findListsInAnnualReview() {
   const today = startOfToday().toISOString();
