+++ src/scheduler/workers/unpublish/weekly/findNonRespondentsForList.ts
@@ -1,4 +1,4 @@
-import { prisma } from "server/models/db/prisma-client";
+import { prisma } from "shared/prisma";;
 import { ListJsonData } from "server/models/types";
 import { schedulerLogger } from "scheduler/logger";
 import { List, Prisma } from "@prisma/client";
