+++ src/server/components/dashboard/annualReview/helpers.keyDates.ts
@@ -1,7 +1,7 @@
 import { addDays, parseISO, subDays } from "date-fns";
 import { ScheduledProcessKeyDates } from "server/models/types";
 import { INTERVALS_IN_DAYS } from "server/config";
-import { prisma } from "server/models/db/prisma-client";
+import { prisma } from "shared/prisma";;
 import { ListWithJsonData } from "server/components/dashboard/helpers";
 import { logger } from "server/services/logger";
