+++ src/server/models/db/__mocks__/prisma-client.ts
@@ -1,6 +1,6 @@
 import { PrismaClient } from "@prisma/client";
 import { mockDeep, mockReset, MockProxy } from "jest-mock-extended";
-import { prisma as prismaClient } from "../prisma-client";
+import { prisma as prismaClient } from "shared/prisma";
 
 jest.mock("../prisma-client", () => ({
   __esModule: true,
