+++ src/server/models/listItem/__tests__/listItem.update.test.ts
@@ -1,6 +1,6 @@
 import * as location from "server/services/location";
 import * as listItem from "./../listItem";
-import { prisma } from "server/models/db/prisma-client";
+import { prisma } from "shared/prisma";;
 
 jest.mock("../../db/prisma-client");
