+++ src/server/models/listItem/listItemEvent.ts
@@ -1,4 +1,4 @@
-import { prisma } from "../db/prisma-client";
+import { prisma } from "shared/prisma";
 import { ListItemEvent, Prisma } from "@prisma/client";
 import { Event, EventCreateInput, EventJsonData } from "./types";
 import { logger } from "server/services/logger";
