+++ src/server/models/listItem/summary.helpers.ts
@@ -4,7 +4,7 @@ import type { ActivityStatusViewModel, AnnualReviewBanner } from "server/models/
 import * as DateFns from "date-fns";
 import { differenceInWeeks, isPast, isWithinInterval, parseISO, set } from "date-fns";
 import type { ListWithJsonData } from "server/components/dashboard/helpers";
-import { prisma } from "server/models/db/prisma-client";
+import { prisma } from "shared/prisma";;
 
 /**
  * Additions to Status type to help with rendering
