+++ src/server/models/listItem/summary.ts
@@ -1,7 +1,7 @@
 import { IndexListItem, ListIndexOptions } from "server/models/listItem/types";
 import { PaginationResults } from "server/components/lists";
 import { calculatePagination, queryToPrismaQueryMap } from "server/models/listItem/queryFactory";
-import { prisma } from "server/models/db/prisma-client";
+import { prisma } from "shared/prisma";;
 import { logger } from "server/services/logger";
 import { getPaginationValues } from "server/models/listItem/pagination";
 import { Prisma } from "@prisma/client";
