+++ src/server/models/user.ts
@@ -1,7 +1,7 @@
 import { omit } from "lodash";
 import { logger } from "server/services/logger";
 import { isGovUKEmailAddress } from "server/utils/validation";
-import { prisma } from "./db/prisma-client";
+import { prisma } from "shared/prisma";
 import { User, UserRoles } from "./types";
 
 export async function findUserByEmail(email: string): Promise<User | undefined> {
